mvc - a work in progress.
Key concepts:

accessors - a single function that can act as
both a getter or setter depending on how it is called.

binding - provides read and update functions that bind a model element to a view element.
If customised with adapters, the types of the view and model values
can be different. Different types of bindings can be used to bind
to different aspects of a view (for example, to bind the the .type property
of an input rather than to the .value property).

model - a map of accessors for an underlying object

views - a map of accessors for underlying DOM elements and FORM INPUTs

controller - a model, a view and a collection of bindings.